+++ changelog.md
@@ -8,7 +8,7 @@
 * [1577](https://github.com/zeta-chain/node/pull/1577) - add chain header tests in E2E tests and fix admin tests
 
 ### Features
-
+* [1658](https://github.com/zeta-chain/node/pull/1658) - modify emission distribution to use fixed block rewards 
 ### Fixes
 * [1535](https://github.com/zeta-chain/node/issues/1535) - Avoid voting on wrong ballots due to false blockNumber in EVM tx receipt
 * [1588](https://github.com/zeta-chain/node/pull/1588) - fix chain params comparison logic